requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
croniter==1.4.1
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-utilities==0.1.3.1
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
loguru==0.7.2
mongomock==4.1.2
numpy==1.26.4
pandas==2.1.4
py-fibonacci==0.5.2
pydantic==2.5.2
pydantic_core==2.14.5
pyi18n-v2==1.2.1
PyJWT==2.8.0
pymongo==4.6.1
pymongo-schema==0.4.1
pytest~=7.2.2
python-dateutil~=2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests~=2.28.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.35.1
testcontainers==4.6.0
typing_extensions~=4.8.0
tzdata==2024.1
uuid==1.30
uvicorn==0.24.0
websocket-client==1.7.0
websockets~=12.0